% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPDBData-package.R
\docType{package}
\name{ChIPDBData}
\alias{ChIPDBData-package}
\alias{ChIPDBData}
\title{ChIPDBData: A package for curated ChIP-seq TF target databases}
\description{
The \code{ChIPDBData} package provides curated ChIP-seq transcription factor
target databases designed for use with TFEA.ChIP
}
\section{Package contents}{

\itemize{
\item Functions provided: \code{getChIPDB} (for loading the dataset)
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yberda/ChIPDBData}
  \item Report bugs at \url{https://github.com/yberda/ChIPDBData/issues}
}

}
\author{
\strong{Maintainer}: Yosra Berrouayel \email{yosraberrouayel@gmail.com} (\href{https://orcid.org/0000-0002-0768-5933}{ORCID})

Authors:
\itemize{
  \item Luis del Peso \email{lpeso@iib.uam.es} (\href{https://orcid.org/0000-0003-4014-5688}{ORCID})
}

}
\keyword{internal}
